feat: add bolus markers and basal area to web dashboard chart
- Add GET /api/integrations/pump/history endpoint exposing pump events
- Transform glucose chart from ScatterChart to ComposedChart (Recharts)
- Overlay basal rate as stepAfter Area with hidden right Y-axis
- Show bolus deliveries as dashed ReferenceLine markers with unit labels
- Add usePumpEvents hook mirroring glucose history fetch pattern
- Fix mobile basal sync gap: backfilled basals now enqueued to backend
- Add enqueueBasalBatch to SyncQueueEnqueuer for batch operations
- Tooltip shows basal rate on hover, glucose value for scatter points
- Legend conditionally shows Bolus/Basal items when data exists

+++ apps/api/src/services/tandem_sync.py
@@ -1006,7 +1006,7 @@ async def get_latest_pump_event(
 async def get_pump_events(
     db: AsyncSession,
     user_id: uuid.UUID,
-    hours: int = 24,
+    hours: float = 24,
     limit: int = 100,
     event_type: PumpEventType | None = None,
 ) -> list[PumpEvent]:
